requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.11
SQLAlchemy==1.4.49
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
opencv-python-headless==4.8.1.76
ffmpeg-python==0.3.0
numpy==1.26.2
Pillow==10.0.0
torch==2.2.0
torchvision==0.16.1
scipy==1.11.3
librosa==0.10.1
matplotlib==3.8.1
aiofiles==23.1.0
typing-extensions==4.7.1
